backend/requirements/production.txt
# Зависимости для продакшена
-r base.txt

# Веб-сервер
gunicorn==21.2.0
whitenoise==6.6.0

# Мониторинг и логирование
sentry-sdk==1.38.0
django-health-check==3.17.0

# Кэширование
django-redis==5.4.0

# Безопасность
django-security==0.18.0
django-csp==3.7

# Производительность
django-cachalot==2.6.1
